+++ src/config.ts
@@ -12,7 +12,9 @@ export default class Config {
   public static deserialize(buffer: StaticDataView): Config {
     return new Config({
       debug: false,
+      enableCompression: buffer.getBool(),
       enableOptimizations: buffer.getBool(),
+      integrityCheck: buffer.getBool(),
       loadCosmeticFilters: buffer.getBool(),
       loadGenericCosmeticsFilters: buffer.getBool(),
       loadNetworkFilters: buffer.getBool(),
@@ -20,27 +22,35 @@ export default class Config {
   }
 
   public readonly debug: boolean;
+  public readonly enableCompression: boolean;
   public readonly enableOptimizations: boolean;
+  public readonly integrityCheck: boolean;
   public readonly loadCosmeticFilters: boolean;
   public readonly loadGenericCosmeticsFilters: boolean;
   public readonly loadNetworkFilters: boolean;
 
   constructor({
     debug = false,
+    enableCompression = false,
     enableOptimizations = true,
+    integrityCheck = true,
     loadCosmeticFilters = true,
     loadGenericCosmeticsFilters = true,
     loadNetworkFilters = true,
   }: Partial<Config> = {}) {
     this.debug = debug;
+    this.enableCompression = enableCompression;
     this.enableOptimizations = enableOptimizations;
+    this.integrityCheck = integrityCheck;
     this.loadCosmeticFilters = loadCosmeticFilters;
     this.loadGenericCosmeticsFilters = loadGenericCosmeticsFilters;
     this.loadNetworkFilters = loadNetworkFilters;
   }
 
   public serialize(buffer: StaticDataView): void {
+    buffer.pushBool(this.enableCompression);
     buffer.pushBool(this.enableOptimizations);
+    buffer.pushBool(this.integrityCheck);
     buffer.pushBool(this.loadCosmeticFilters);
     buffer.pushBool(this.loadGenericCosmeticsFilters);
     buffer.pushBool(this.loadNetworkFilters);